+++ apps/user-office-backend/src/asyncJobs/allAsyncJobs.ts
@@ -1,14 +1,12 @@
-import checkCallsEndedJob from './jobs/checkCallsEnded';
-import checkCallsEndedInternalJob from './jobs/checkCallsEndedInternal';
+import checkAllCallsEndedJob from './jobs/checkAllCallsEnded';
 import checkCallsReviewEndedJob from './jobs/checkCallsReviewEnded';
 import checkCallsSEPReviewEndedJob from './jobs/checkCallsSEPReviewEnded';
 import { UserOfficeAsyncJob } from './startAsyncJobs';
 
 const ALL_ASYNC_JOBS: UserOfficeAsyncJob[] = [
-  checkCallsEndedJob,
+  checkAllCallsEndedJob,
   checkCallsReviewEndedJob,
   checkCallsSEPReviewEndedJob,
-  checkCallsEndedInternalJob,
 ];
 
 export default ALL_ASYNC_JOBS;
