--- apps/user-office-backend/src/asyncJobs/jobs/checkCallsEndedInternal.ts
@@ -1,59 +1,0 @@
-import { logger } from '@user-office-software/duo-logger';
-
-import { CallDataSource } from '../../datasources/CallDataSource';
-import { eventBus } from '../../events';
-import { Event } from '../../events/event.enum';
-import { UserOfficeAsyncJob } from '../startAsyncJobs';
-
-const checkCallsEndedInternal = async (dataSource: CallDataSource) => {
-  const isTestingMode = process.env.NODE_ENV === 'test';
-
-  try {
-    const notEndedInternalCalls = await dataSource.getCalls({
-      isEndedInternal: false,
-    });
-
-    const currentDate = new Date();
-
-    const callsThatShouldEndInternal = notEndedInternalCalls.filter(
-      (notEndedCall) =>
-        notEndedCall.endCallInternal.getTime() <= currentDate.getTime()
-    );
-
-    const updatedCalls = [];
-
-    for (const callThatShouldEndInternal of callsThatShouldEndInternal) {
-      const updatedCall = await dataSource.update({
-        ...callThatShouldEndInternal,
-        callEndedInternal: true,
-      });
-
-      if (!isTestingMode) {
-        // NOTE: Fire the "CALL_ENDED" event if not in testing mode.
-        eventBus.publish({
-          type: Event.CALL_ENDED_INTERNAL,
-          call: updatedCall,
-          isRejection: false,
-          key: 'call',
-          loggedInUserId: 0,
-        });
-      }
-
-      updatedCalls.push(updatedCall);
-    }
-
-    return updatedCalls;
-  } catch (error) {
-    logger.logException('Checking and ending calls internal failed: ', error);
-  }
-};
-
-// NOTE: Run each hour at minute 0
-const options = { timeToRun: '0 * * * *' };
-
-const checkCallsEndedInternalJob: UserOfficeAsyncJob = {
-  functionToRun: checkCallsEndedInternal,
-  options,
-};
-
-export default checkCallsEndedInternalJob;
